mood_tracker_backend/requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==4.2.28
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.11
psycopg2-binary==2.9.10
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.13.2
urllib3==2.6.3
